% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.mcmc.R
\name{tidy.mcmc}
\alias{tidy.mcmc}
\title{tidy.mcmc}
\usage{
tidy.mcmc(x, q = c(0.025, 0.5, 0.975))
}
\arguments{
\item{x}{an MCMC list from, e.g. coda.samples()}

\item{q}{a list of quantiles, such as those corresponding to a 95\% credible and its median}
}
\description{
A function that behaves like those from "broom", tidy.mcmc will take an mcmc.list object from coda.samples and return a data frame that summarises each parameters with its mean and quantiles and returns the output as a data.frame object. It does not yet work as though it were a function in broom, e.g. tidy(an.mcmc.list) doesn't work
}

